# backend/src/dtos/transaction_dto.py
from pydantic import BaseModel
from datetime import date
from decimal import Decimal
from typing import Optional


class TransactionBase(BaseModel):
    title: str
    description: Optional[str] = None
    amount: Decimal
    transaction_date: date
    transaction_type: str
    category_id: str


class TransactionCreate(TransactionBase):
    pass


class TransactionDTO(TransactionBase):
    id: str

    class Config:
        from_attributes = True